# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.EndWorkoutPayload:
    properties:
      duration:
        type: string
      workout_id:
        type: integer
    type: object
...
  handlers.WorkoutReviewPayload:
    properties:
      content:
        type: string
      rating:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      title:
        maxLength: 255
        type: string
    required:
    - content
    - rating
    - title
    type: object
...
  store.Exercise:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: integer
      is_duration:
        type: boolean
      like:
        type: integer
      muscles:
        items:
          type: string
        type: array
      name:
        type: string
      tutorial_link:
        type: string
      user_id:
        type: integer
    type: object
...
  store.WorkoutReview:
    properties:
      content:
        type: string
      created_at:
        type: string
      rating:
        type: integer
      title:
        type: string
      user_id:
        type: integer
      workout_id:
        type: integer
    type: object
  store.WorkoutReviewWithMetadata:
    properties:
      content:
        type: string
      created_at:
        type: string
      rating:
        type: integer
      title:
        type: string
      user_id:
        type: integer
      username:
        type: string
      workout_id:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Atom Fit API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Atom Fit API
paths:
...
  /exercises:
    get:
      consumes:
      - application/json
      description: Get all Exercises
      parameters:
      - description: Since
        in: query
        name: since
        type: string
      - description: Until
        in: query
        name: until
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Tags
        in: query
        name: tags
        type: string
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Exercise'
            type: array
      summary: Get all Exercises
      tags:
      - exercises
    post:
      consumes:
      - application/json
      description: Create a new Exercise
      parameters:
      - description: Exercise Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/store.Exercise'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Exercise'
      security:
      - ApiKeyAuth: []
      summary: Create a new Exercise
      tags:
      - exercises
...
  /exercises/{id}:
    get:
      consumes:
      - application/json
      description: Get Exercise by id from param
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Exercise'
      summary: Get Exercise by id from param
      tags:
      - exercises
  /exercises/{idString}:
    delete:
      consumes:
      - application/json
      description: Delete Exercise by idString
      parameters:
      - description: Exercise ID
        in: path
        name: idString
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/store.Exercise'
      security:
      - ApiKeyAuth: []
      summary: Delete Exercise by idString
      tags:
      - exercises
    patch:
      consumes:
      - application/json
      description: Update Exercise by idString
      parameters:
      - description: Exercise ID
        in: path
        name: idString
        required: true
        type: integer
      - description: Exercise Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/store.Exercise'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Exercise'
      security:
      - ApiKeyAuth: []
      summary: Update Exercise by idString
      tags:
      - exercises
  /exercises/{userID}:
    get:
      consumes:
      - application/json
      description: Get all Exercises by user id
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Exercise'
            type: array
      summary: Get all Exercises by user id
      tags:
      - exercises
...
  /reviews/workout/{workoutID}:
    get:
      consumes:
      - application/json
      description: Get workout reviews
      parameters:
      - description: Workout ID
        in: path
        name: workoutID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.WorkoutReviewWithMetadata'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get workout reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Review workout
      parameters:
      - description: Workout ID
        in: path
        name: workoutID
        required: true
        type: integer
      - description: Review workout payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.WorkoutReviewPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.WorkoutReview'
        "400":
          description: Bad Request
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Review workout
      tags:
      - reviews
  /reviews/workout/{workoutID}/{reviewID}:
    delete:
      consumes:
      - application/json
      description: Delete workout review
      parameters:
      - description: Workout ID
        in: path
        name: workoutID
        required: true
        type: integer
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/store.WorkoutReview'
      security:
      - ApiKeyAuth: []
      summary: Delete workout review
      tags:
      - reviews
    patch:
      consumes:
      - application/json
      description: Update workout review
      parameters:
      - description: Workout ID
        in: path
        name: workoutID
        required: true
        type: integer
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: integer
      - description: Review workout payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.WorkoutReviewPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.WorkoutReview'
        "400":
          description: Bad Request
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update workout review
      tags:
      - reviews
...
  /workouts/end:
    post:
      consumes:
      - application/json
      description: End workout
      parameters:
      - description: End workout payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.EndWorkoutPayload'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: End workout
      tags:
      - finished_workouts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"